LayoutTests/platform/mac/accessibility/element-haspopup-expected.txt

Link 1 Link 2
This tests that an element will report that it has a popup correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.hasPopup is true
PASS accessibilityController.focusedElement.hasPopup is false
PASS successfullyParsed is true

TEST COMPLETE
